chore: update getData test
The `getData` test is renamed `strapiDataProvider` test and it only
tests the Strapi implementation.

TODO: The current version of the test only covers one of the
`DataProvider` methods (`getAllParties`) and does not even test all of
its options. The coverage should be extended in the future.
