--- frontend/tests/data/locales.json
@@ -1,8 +1,0 @@
-[
-  {
-    "id": 1,
-    "code": "en",
-    "isDefault": true,
-    "name": "English"
-  }
-]
