--- frontend/tests/data/singleParty.json
@@ -1,36 +1,0 @@
-{
-  "data": {
-    "id": 3370,
-    "attributes": {
-      "name": "Mature Configuration Somewhere",
-      "info": "Officia ratione quas. Blanditiis doloremque voluptatibus nobis. Rem at atque corporis ipsam nihil aspernatur optio id.",
-      "shortName": "MCS",
-      "mainColor": "Green",
-      "createdAt": "2023-05-06T09:38:08.841Z",
-      "updatedAt": "2023-05-06T09:38:08.862Z",
-      "publishedAt": "2023-05-06T09:38:08.838Z",
-      "locale": "en",
-      "logo": {
-        "data": null
-      },
-      "localizations": {
-        "data": [
-          {
-            "id": 3381,
-            "attributes": {
-              "name": "Mature Configuration Somewhere",
-              "info": "Minus facere rerum fugiat exercitationem tenetur iusto. Facilis quibusdam distinctio alias incidunt soluta. Corrupti eos accusantium atque eius.",
-              "shortName": "MCS",
-              "mainColor": "Green",
-              "createdAt": "2023-05-06T09:38:08.841Z",
-              "updatedAt": "2023-05-06T09:38:08.906Z",
-              "publishedAt": "2023-05-06T09:38:08.845Z",
-              "locale": "fi-FI"
-            }
-          }
-        ]
-      }
-    }
-  },
-  "meta": {}
-}
